fixed some broken usage of select{}
I assumed that select{} would always pick the first case that can yield
when multiple channels can yield, but the case is instead chosen
pseudo-randomly.

this caused drop_chan to malfunction: when quitting there would be a
50/50 chance that new messages would get queued rather than dropped as
intended.

it also caused queue consumption to act randomly when quitting; if the
queue was non-empty then each goroutine in the process pool would
randomly either quit or consume another queue entry. the desired
behavior is of course that no goroutine quits before the queue is empty.
this is now accomplished by closing the channel which works nicely as a
broadcasted signal which cannot be "seen" before the queue is empty.